+++ CHANGELOG.md
@@ -37,3 +37,4 @@
 36. Render Filters on Profile Page using React component
 37. Render SessionSummary on Session Page & Profile Page using Component
 38. Improve Table Handling in SessionSummary Component
+39. Render TargetList using React Component